+++ DataExplorer/src/gde/utils/CurveUtils.java
@@ -182,7 +182,7 @@ public class CurveUtils {
 		int recordSize = isCompareSet ? record.realSize() : record.size();
 
 		// calculate time line adaption if record set is compare set, compare set max have different times for each record, (intRecordSize - 1) is number of time deltas for calculation 
-		int timeStep = record.getTimeStep_ms();
+		double timeStep = record.getTimeStep_ms();
 		double adaptXMaxValue = isCompareSet ? (1.0 * (recordSize - 1) * record.getParent().getSize() / (recordSize - 1) * timeStep) : (1.0 * (recordSize - 1) * timeStep);
 		if (CurveUtils.log.isLoggable(Level.FINE)) CurveUtils.log.fine("recordSize = " + recordSize + " adaptXMaxValue = " + adaptXMaxValue);
